+++ parser_library/include/analyzer.h
@@ -27,6 +27,7 @@
 
 #include "analyzing_context.h"
 #include "compiler_options.h"
+#include "external_functions.h"
 #include "preprocessor_options.h"
 #include "processing_format.h"
 #include "protocol.h"
@@ -113,6 +114,7 @@ class analyzer_options
     std::string dep_name;
     processing::processing_kind dep_kind = processing::processing_kind::ORDINARY;
     diagnostic_limit diag_limit;
+    external_functions_list external_functions;
 
     void set(utils::resource::resource_location rl) { file_loc = std::move(rl); }
     void set(parse_lib_provider* lp) { lib_provider = lp; }
@@ -132,6 +134,7 @@ class analyzer_options
         dep_name = std::move(d.name);
     }
     void set(diagnostic_limit dl) { diag_limit = dl; }
+    void set(external_functions_list ef) { external_functions = std::move(ef); }
 
     context::hlasm_context& get_hlasm_context();
     analyzing_context& get_context();
@@ -167,8 +170,9 @@ public:
         constexpr auto o_cnt = (0 + ... + std::is_convertible_v<std::decay_t<Args>, output_handler*>);
         constexpr auto dep_data_cnt = (0 + ... + std::is_convertible_v<std::decay_t<Args>, dependency_data>);
         constexpr auto diag_limit_cnt = (0 + ... + std::is_convertible_v<std::decay_t<Args>, diagnostic_limit>);
+        constexpr auto ef_cnt = (0 + ... + std::is_same_v<std::decay_t<Args>, external_functions_list>);
         constexpr auto cnt = rl_cnt + lib_cnt + ao_cnt + ac_cnt + hi_cnt + f_oc_cnt + ids_cnt + pp_cnt + vfm_cnt
-            + fmc_cnt + o_cnt + dep_data_cnt + diag_limit_cnt;
+            + fmc_cnt + o_cnt + dep_data_cnt + diag_limit_cnt + ef_cnt;
 
         static_assert(rl_cnt <= 1, "Duplicate resource_location");
         static_assert(lib_cnt <= 1, "Duplicate parse_lib_provider");
@@ -180,11 +184,13 @@ public:
         static_assert(pp_cnt <= 1, "Duplicate preprocessor_args");
         static_assert(vfm_cnt <= 1, "Duplicate virtual_file_monitor");
         static_assert(fmc_cnt <= 1, "Duplicate fade message container");
-        static_assert(!(ac_cnt && (ao_cnt || ids_cnt || pp_cnt)),
-            "Do not specify both analyzing_context and asm_option, id_storage or preprocessor_args");
+        static_assert(!(ac_cnt && (ao_cnt || ids_cnt || pp_cnt || ef_cnt)),
+            "Do not specify both analyzing_context and asm_option, id_storage, preprocessor_args or "
+            "external_functions");
         static_assert(o_cnt <= 1, "Duplicate output_handler");
         static_assert(dep_data_cnt <= 1, "Duplicate dependency_data");
         static_assert(diag_limit_cnt <= 1, "Duplicate diagnostic_limit");
+        static_assert(ef_cnt <= 1, "Duplicate external_functions");
         static_assert(cnt == sizeof...(Args), "Unrecognized argument provided");
 
         (set(std::forward<Args>(args)), ...);