+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## ****Unreleased****
 
+#### Added
+- SETAF and SETCF instruction support
+
 ## [1.20.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.19.0...1.20.0) (2025-11-12)
 
 #### Added